requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
regex>=2023.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning
torch>=2.0.0

# Voice Recognition (optional)
SpeechRecognition>=3.10.0
# Pour le support microphone (optionnel): pip install pyaudio
# Pour Whisper local (optionnel): pip install openai-whisper
